Adds to_string block with optional types
Duplicate commit dac4d63 from godot-block-coding by DoomTas3r.

The to_string block converts a parameter to a string. It includes an option input with an integer, a boolean, a vector2, a vector3, a color, and an object.

The type of the selected input gets detected and displays as a normal parameter input with the detected type and snapping.